pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "card_detector"
version = "0.1.0"
description = "YOLO → CNN card detection pipeline"
readme = "README.md"
requires-python = ">=3.8"
authors = [{ name = "OctalDecoder", email = "octaldecoder@gmail.com" }]
dependencies = [
    "pyyaml",
    "Pillow",
    "numpy",
    "torch",
    "torchvision",
    "albumentations",
    "ultralytics",
    "opencv-python>=4.8",
]

[project.scripts]
card-detector = "card_detector.__main__:main"
train-cnn = "card_detector.cnn.train:main"

[tool.setuptools.packages.find]
where = ["src"]
